+++ log/log_test.go
@@ -20,7 +20,15 @@ type LogMiddlewareTestSuite struct {
 
 type testWriter struct {
 	io.Writer
-	closed bool
+	closed     bool
+	preWritten bool
+}
+
+func (w *testWriter) PreWrite(b []byte) {
+	w.preWritten = true
+	if pr, ok := w.Writer.(goyave.PreWriter); ok {
+		pr.PreWrite(b)
+	}
 }
 
 func (w *testWriter) Close() error {
@@ -118,6 +126,7 @@ func (suite *LogMiddlewareTestSuite) TestCloseChildWriter() {
 		}
 		resp.Body.Close()
 		suite.True(closeableWriter.closed)
+		suite.True(closeableWriter.preWritten)
 	})
 }